patch character classes

+++ src/main/java/longroad/annemie/PathfinderAPI/Skill/SkillRepository.java
@@ -2,6 +2,6 @@ package longroad.annemie.PathfinderAPI.Skill;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SkillRepository extends JpaRepository<Skill, Integer>
+public interface SkillRepository extends JpaRepository<Skill, Short>
 {
 }
